drivers: thermal: step_wise: Handle the temperature stable trend
Step wise algorithm ignores the temperature stable trend, which can
cause a case where step wise will not clear the mitigation, when the
temperature is lower than hysteresis and stays stable throughout.

In order to avoid this, consider temperature stable trend similar to temperature dropping trend. This will lower the mitigation when the temperature is less than hysteresis.